requirements.txt

click==8.0.3
colorama==0.4.4
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.30
debugpy==1.5.1

Django==4.0.3
docutils==0.18.1
entrypoints==0.4
ephem==4.1.3
executing==0.8.3
filelock==3.6.0
Flask==2.0.2
fonttools==4.32.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
hijri-converter==2.2.3

image==1.5.33
importlib-metadata==4.11.3
IOTk==1.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0


MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
MultiApp==1.0
multitasking==0.0.10
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pafy==0.5.5
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.4.8
platformdirs==2.5.1
playsound==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
streamlit==1.7.0
yfinance==0.1.70
prophet==1.1.4
